Make BeatmapSetOverlay use OverlayHeader
